Add files via upload
This dataset consists of TV shows and movies available on Netflix as of 2019. In 2018, they released an interesting report that shows that the number of TV shows on Netflix has nearly tripled since 2010. The number of movies has decreased by more than 2,000 titles since 2010, while the number of TV shows has almost tripled. It will be interesting to explore what other insights can be obtained from the same dataset.